+++ test/document.test.js
@@ -1,8 +1,15 @@
 import {expect} from 'chai';
 import {Document} from '../lib/document.js';
 import {fail, validateId, expectError, Data} from './util.js';
-import {IS_BASE_DOCUMENT, IS_DOCUMENT, IS_EMBEDDED, isDocument, isEmbeddedDocument, isReferenceable} from '../lib/validate.js';
-import {ValidationError} from '../lib/errors.js';
+import {
+    IS_BASE_DOCUMENT,
+    IS_DOCUMENT,
+    IS_EMBEDDED,
+    isDocument,
+    isEmbeddedDocument,
+    isReferenceable,
+    ValidationError
+} from '../lib/validate.js';
 import {initMochaHooksForNedb} from './database.js';